Implement up scaling for edge weights
Implemented a basic version of edges increasing in size when the weight
increases. Still need to code some edge case handling for nodes going
off screen when edges increase in size. Some duplicate code can be
removed and need to implement a function to updates positions of
pre-existing edges when a node is moved.